much refactoring
