+++ cromwell-drs-localizer/src/main/scala/drs/localizer/downloaders/Downloader.scala
@@ -1,0 +1,9 @@
+package drs.localizer.downloaders
+
+import cats.effect.{ExitCode, IO}
+
+trait Downloader {
+
+  def download: IO[ExitCode]
+
+}
